+++ app/victoria-traces/servicegraph/servicegraph.go
@@ -16,11 +16,12 @@ import (
 )
 
 var (
-	enableServiceGraphTask     = flag.Bool("servicegraph.enableTask", false, "Whether to enable background task for generating service graph. It should only be enabled on VictoriaTraces single-node or vtstorage.")
-	serviceGraphTaskInterval   = flag.Duration("servicegraph.taskInterval", time.Minute, "The background task interval for generating service graph data. It requires setting -servicegraph.enableTask=true.")
-	serviceGraphTaskTimeout    = flag.Duration("servicegraph.taskTimeout", 30*time.Second, "The background task timeout duration for generating service graph data. It requires setting -servicegraph.enableTask=true.")
-	serviceGraphTaskLookbehind = flag.Duration("servicegraph.taskLookbehind", time.Minute, "The lookbehind window for each time service graph background task run. It requires setting -servicegraph.enableTask=true.")
-	serviceGraphTaskLimit      = flag.Uint64("servicegraph.taskLimit", 1000, "How many service graph relations each task could fetch for each tenant. It requires setting -servicegraph.enableTask=true.")
+	enableServiceGraphTask        = flag.Bool("servicegraph.enableTask", false, "Whether to enable background task for generating service graph. It should only be enabled on VictoriaTraces single-node or vtstorage.")
+	serviceGraphTaskInterval      = flag.Duration("servicegraph.taskInterval", time.Minute, "The background task interval for generating service graph data. It requires setting -servicegraph.enableTask=true.")
+	serviceGraphTaskTimeout       = flag.Duration("servicegraph.taskTimeout", 30*time.Second, "The background task timeout duration for generating service graph data. It requires setting -servicegraph.enableTask=true.")
+	serviceGraphTaskLookbehind    = flag.Duration("servicegraph.taskLookbehind", time.Minute, "The lookbehind window for each time service graph background task run. It requires setting -servicegraph.enableTask=true.")
+	serviceGraphTaskLimit         = flag.Uint64("servicegraph.taskLimit", 1000, "How many service graph relations each task could fetch for each tenant. It requires setting -servicegraph.enableTask=true.")
+	serviceGraphDatabaseTaskLimit = flag.Uint64("servicegraph.databaseTaskLimit", 1000, "How many service-database graph relations each task could fetch for each tenant. It requires setting -servicegraph.enableTask=true. Setting it to 0 will stop VictoriaTraces from generating service-database graph relations.")
 )
 
 var (
@@ -97,25 +98,40 @@ func GenerateServiceGraphTimeRange(ctx context.Context) {
 	}
 	commonFieldLen := len(commonFields)
 
-	var hasError bool
+	var (
+		rows     [][]logstorage.Field
+		hasError bool
+	)
 	// query and persist operations are executed sequentially, which helps not to consume excessive resources.
 	for _, tenantID := range tenantIDs {
-		// query service graph relations
-		rows, err := vtselect.GetServiceGraphTimeRange(ctx, tenantID, startTime, endTime, *serviceGraphTaskLimit)
+		// query service-to-service graph relations
+		rows, err = vtselect.GetServiceGraphTimeRange(ctx, tenantID, startTime, endTime, *serviceGraphTaskLimit)
 		if err != nil {
 			hasError = true
 			logger.Errorf("cannot get service graph for time range [%d, %d] of tenant %s: %s", startTime.Unix(), endTime.Unix(), tenantID.String(), err)
 			continue
 		}
-		if len(rows) == 0 {
-			continue
+
+		// query service-to-db relations if flag is set/defaulted to non-zero value.
+		if *serviceGraphDatabaseTaskLimit > 0 {
+			dbRows, err := vtselect.GetServiceDBGraphTimeRange(ctx, tenantID, startTime, endTime, *serviceGraphDatabaseTaskLimit)
+			if err != nil {
+				hasError = true
+				logger.Errorf("cannot get middleware graph for time range [%d, %d] of tenant %s: %s", startTime.Unix(), endTime.Unix(), tenantID.String(), err)
+				continue
+			}
+			rows = append(rows, dbRows...)
 		}
-		commonFields = commonFields[:commonFieldLen]
-		// persist service graph relations
-		commonFields, err = vtinsert.PersistServiceGraph(ctx, tenantID, commonFields, rows, endTime)
-		if err != nil {
-			hasError = true
-			logger.Errorf("cannot persist service graph for time range [%d, %d] of tenant %s: %s", startTime.Unix(), endTime.Unix(), tenantID.String(), err)
+
+		if len(rows) > 0 {
+			commonFields = commonFields[:commonFieldLen]
+			// persist service graph relations
+			commonFields, err = vtinsert.PersistServiceGraph(ctx, tenantID, commonFields, rows, endTime)
+			if err != nil {
+				hasError = true
+				logger.Errorf("cannot persist service graph for time range [%d, %d] of tenant %s: %s", startTime.Unix(), endTime.Unix(), tenantID.String(), err)
+				continue
+			}
 		}
 	}